+++ src/index.ts
@@ -35,6 +35,7 @@ interface GitHubOptions {
 export interface GitHubReleaseOptions {
   releaseLabel?: string;
   draft?: boolean;
+  skipGithubRelease?: boolean;
 }
 
 // Used by ReleasePR: Factory and Constructor
@@ -54,6 +55,7 @@ export interface ReleasePROptions {
   // and then remove this property
   versionFile?: string;
   pullRequestTitlePattern?: string;
+  signoff?: string;
   extraFiles?: string[];
 }
 
@@ -71,6 +73,7 @@ interface ReleaserConstructorOptions {
 interface ManifestOptions {
   configFile?: string;
   manifestFile?: string;
+  signoff?: string;
 }
 
 export interface ManifestConstructorOptions
@@ -101,6 +104,7 @@ export type ManifestPackage = Pick<
   | 'releaseAs'
   | 'changelogSections'
   | 'changelogPath'
+  | 'skipGithubRelease'
 > & {
   // these items are not optional in the manifest context.
   path: string;