+++ src/manifest.ts
@@ -42,6 +42,7 @@ import {ReleasePR} from './release-pr';
 import {Changes} from 'code-suggester';
 import {PluginType, getPlugin} from './plugins';
 import {ManifestPlugin} from './plugins/plugin';
+import {signoffCommitMessage} from './util/signoff-commit-message';
 
 interface ReleaserConfigJson {
   'release-type'?: ReleaseType;
@@ -49,6 +50,7 @@ interface ReleaserConfigJson {
   'bump-patch-for-minor-pre-major'?: boolean;
   'changelog-sections'?: ChangelogSection[];
   'release-as'?: string;
+  'skip-github-release'?: boolean;
   draft?: boolean;
 }
 
@@ -85,12 +87,14 @@ export class Manifest {
   checkpoint: Checkpoint;
   configFile?: Config;
   headManifest?: ManifestJson;
+  signoff?: string;
 
   constructor(options: ManifestConstructorOptions) {
     this.gh = options.github;
     this.configFileName = options.configFile || RELEASE_PLEASE_CONFIG;
     this.manifestFileName = options.manifestFile || RELEASE_PLEASE_MANIFEST;
     this.checkpoint = options.checkpoint || checkpoint;
+    this.signoff = options.signoff;
   }
 
   protected async getBranchName() {
@@ -241,6 +245,10 @@ export class Manifest {
             config['release-as']
           ),
           draft: pkgCfg['draft'] ?? config['draft'],
+          skipGithubRelease:
+            pkgCfg['skip-github-release'] ??
+            config['skip-github-release'] ??
+            false,
         };
         packages.push(pkg);
       }
@@ -631,9 +639,18 @@ export class Manifest {
       newManifestVersions,
       pkgsWithChanges
     );
+
+    const title = `chore: release ${await this.gh.getDefaultBranch()}`;
+
+    // Sign-off message if signoff option is enabled
+    const message = this.signoff
+      ? signoffCommitMessage(title, this.signoff)
+      : title;
+
     const pr = await this.gh.openPR({
       branch: branchName,
-      title: `chore: release ${await this.gh.getDefaultBranch()}`,
+      title,
+      message,
       body: body,
       updates: [],
       labels: DEFAULT_LABELS,
@@ -698,6 +715,14 @@ export class Manifest {
         this.checkpoint(
           `Unable to find last version for ${pkgLogDisp}.`,
           CheckpointType.Failure
+        );
+        releases[pkg.config.path] = undefined;
+        continue;
+      }
+      if (pkg.config.skipGithubRelease) {
+        this.gh.commentOnIssue(
+          `:robot: ${pkgName} not configured for release :no_entry_sign:`,
+          lastMergedPR.number
         );
         releases[pkg.config.path] = undefined;
         continue;